Organized Classes into Packages according to their Component
