--- src/main/java/cs2103/v15_1j/jimjim/Parser.java
@@ -1,5 +1,0 @@
-package cs2103.v15_1j.jimjim;
-
-public interface Parser {
-	public Command parse(String userCommand);
-}
