+++ src/main/java/cs2103/v15_1j/jimjim/command/AddEventCommand.java
@@ -1,8 +1,10 @@
-package cs2103.v15_1j.jimjim;
+package cs2103.v15_1j.jimjim.command;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import cs2103.v15_1j.jimjim.model.TaskEvent;
+import cs2103.v15_1j.jimjim.searcher.Searcher;
+import cs2103.v15_1j.jimjim.storage.Storage;
 import cs2103.v15_1j.jimjim.model.Event;
 
 public class AddEventCommand implements Command {
