+++ src/main/java/cs2103/v15_1j/jimjim/command/InvalidCommand.java
@@ -1,7 +1,10 @@
-package cs2103.v15_1j.jimjim;
+package cs2103.v15_1j.jimjim.command;
 
 import java.util.List;
+
 import cs2103.v15_1j.jimjim.model.TaskEvent;
+import cs2103.v15_1j.jimjim.searcher.Searcher;
+import cs2103.v15_1j.jimjim.storage.Storage;
 public class InvalidCommand implements Command {
 	
 	private String message;
