+++ src/main/java/cs2103/v15_1j/jimjim/command/MarkDoneCommand.java
@@ -1,9 +1,11 @@
-package cs2103.v15_1j.jimjim;
+package cs2103.v15_1j.jimjim.command;
 
 import java.util.List;
 
 import cs2103.v15_1j.jimjim.model.Task;
 import cs2103.v15_1j.jimjim.model.TaskEvent;
+import cs2103.v15_1j.jimjim.searcher.Searcher;
+import cs2103.v15_1j.jimjim.storage.Storage;
 
 public class MarkDoneCommand implements Command {
     private int taskNum;
