+++ src/main/java/cs2103/v15_1j/jimjim/controller/Controller.java
@@ -1,6 +1,10 @@
-package cs2103.v15_1j.jimjim;
+package cs2103.v15_1j.jimjim.controller;
 import java.util.List;
+
 import cs2103.v15_1j.jimjim.model.TaskEvent;
+import cs2103.v15_1j.jimjim.parser.Parser;
+import cs2103.v15_1j.jimjim.searcher.Searcher;
+import cs2103.v15_1j.jimjim.storage.Storage;
 
 public interface Controller {
 	public String execute(String userCommand);
