+++ src/main/java/cs2103/v15_1j/jimjim/controller/JJController.java
@@ -1,8 +1,13 @@
-package cs2103.v15_1j.jimjim;
+package cs2103.v15_1j.jimjim.controller;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import cs2103.v15_1j.jimjim.command.Command;
 import cs2103.v15_1j.jimjim.model.TaskEvent;
+import cs2103.v15_1j.jimjim.parser.Parser;
+import cs2103.v15_1j.jimjim.searcher.Searcher;
+import cs2103.v15_1j.jimjim.storage.Storage;
 
 public class JJController implements Controller {
