+++ src/main/java/cs2103/v15_1j/jimjim/parser/Parser.java
@@ -1,0 +1,7 @@
+package cs2103.v15_1j.jimjim.parser;
+
+import cs2103.v15_1j.jimjim.command.Command;
+
+public interface Parser {
+	public Command parse(String userCommand);
+}
