+++ src/main/java/cs2103/v15_1j/jimjim/searcher/Searcher.java
@@ -1,4 +1,4 @@
-package cs2103.v15_1j.jimjim;
+package cs2103.v15_1j.jimjim.searcher;
 import java.util.List;
 import cs2103.v15_1j.jimjim.model.TaskEvent;
