+++ src/main/java/cs2103/v15_1j/jimjim/storage/Storage.java
@@ -1,4 +1,4 @@
-package cs2103.v15_1j.jimjim;
+package cs2103.v15_1j.jimjim.storage;
 import java.io.IOException;
 import java.util.List;
 import cs2103.v15_1j.jimjim.model.TaskEvent;
