+++ src/main/java/cs2103/v15_1j/jimjim/ui/UI.java
@@ -1,5 +1,6 @@
-package cs2103.v15_1j.jimjim;
+package cs2103.v15_1j.jimjim.ui;
 
+import cs2103.v15_1j.jimjim.controller.Controller;
 import javafx.stage.Stage;
 
 public interface UI {
