+++ src/test/java/cs2103/v15_1j/jimjim/AddTaskCommandTest.java
@@ -7,6 +7,8 @@ import java.util.ArrayList;
 
 import org.junit.Before;
 import org.junit.Test;
+
+import cs2103.v15_1j.jimjim.command.AddTaskCommand;
 import cs2103.v15_1j.jimjim.model.Task;
 import cs2103.v15_1j.jimjim.model.TaskEvent;
